+++ docs/changelog.md
@@ -8,6 +8,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 Each revision is versioned by the date of the revision.
 
+## 2026-09-2
+
+### Added
+
+- Added support for `strip-prefix` that is passed from the `ingress` relation.
+
 ## 2026-08-20
 
 ### Added